score/requirements.txt
Flask>=2.0
python-dotenv
google-generativeai
scikit-learn
numpy
requests
pandas
rouge-score
gunicorn
psycopg2-binary